refactor(nodes): :art: separate each node into a file
