+++ index.js
@@ -1,8 +1,8 @@
 const { Engine } = require("./src/engine/engine");
 const { Cockpit } = require("./src/cockpit/cockpit");
-const Nodes = require("./src/core/workflow/nodes.js");
-const { nodefyFunction,
-        nodefyClass } = require("./src/core/utils/nodefy");
+
+const Nodes = require("./src/core/workflow/nodes/index.js");
+const { nodefyFunction, nodefyClass } = require("./src/core/utils/nodefy");
 
 module.exports = {
   Engine: Engine,
@@ -10,6 +10,6 @@ module.exports = {
   Nodes: Nodes,
   NodeUtils: {
     nodefyFunction: nodefyFunction,
-    nodefyClass: nodefyClass
-  }
+    nodefyClass: nodefyClass,
+  },
 };
