+++ src/core/utils/node_factory.js
@@ -1,5 +1,5 @@
 /* eslint-disable indent */
-const nodes = require("../workflow/nodes");
+const nodes = require("../workflow/nodes/index.js");
 
 const extra_system_category_map = {};
 function getServiceTask(node_spec) {
